emg: Add hand gesture recognition
This commit adds a HDC-based hand gesture recognition model based on the
paper "Hyperdimensional biosignal processing: A case study for EMG-based
hand gesture recognition". The original code used in the paper is
available as a matlab script at https://github.com/abbas-rahimi/HDC-EMG.

The emg program implements a reduced version of their original matlab
script. The main difference is that the emg uses binary hyper vectors
and is capable of only producing the results for Figures 5 and 7 in the
paper. Moreover, emg is only capable of generating datasets 'inorder'
and does not support the 'random' dataset generation available in
genTrainData() function (please refer to the matlab script).

The emg achieves accuracy similar to those showed in the
aforementioned Figures, i.e., for the spatial encoder and the temporal
encoder using N-grams = 4.