voicehd: Add options to save and load trained models
Add --save-model to save a trained model to a file, and --load-model to
load it. This commit enables training and inference only sessions. The
argument given to --save-model and --load-model must be directory path.
In the first case, the directory must exist, whereas in the second it
must contain the im.txt, cim.txt, and am.txt files.